Разбивает текущий код на модули
